Added createRouteGroup helper
